backend/atomic-services/swap-constraints/requirements.txt
flask
flask_sqlalchemy
marshmallow
python-dotenv
sqlalchemy
psycopg2-binary
